moved remove_star to journalist_app/col.py
